Conflicts:
	src/ServiceControl.AcceptanceTests/AcceptanceTest.cs

+++ src/ServiceControl.AcceptanceTests/AcceptanceTest.cs
@@ -24,6 +24,8 @@
     [TestFixture]
     public abstract class AcceptanceTest
     {
+        protected Dictionary<string, string> AppConfigurationSettings = new Dictionary<string, string>();
+
         public AcceptanceTest()
         {
         }
@@ -81,9 +83,7 @@
                 transportToUse = GetTransportIntegrationFromEnvironmentVar();
             }
 
-            pathToAppConfig = Path.Combine(Path.GetTempPath(), Path.GetTempFileName());
-            InitialiseAppConfig();
-
+       
             Console.Out.WriteLine("Using transport " + transportToUse.Name);
 
             Conventions.EndpointNamingConvention = t =>
@@ -100,11 +100,18 @@
         public void Cleanup()
         {
             Delete(ravenPath);
-            File.Delete(pathToAppConfig);
-            transportToUse.TearDown();
+            
+            if (!string.IsNullOrWhiteSpace(pathToAppConfig))
+            {
+                File.Delete(pathToAppConfig);
+            }
+
+			transportToUse.TearDown();
         }
 
         string ravenPath;
+
+        
 
         static int FindAvailablePort(int startPort)
         {
@@ -165,6 +172,12 @@
             else
             {
                 appSettingsElement.Add(new XElement("add", new XAttribute("key", "ServiceControl/CreateIndexSync"), new XAttribute("value", true)));
+            }
+
+            // Mash any user defined settings into the config
+            foreach (var configSetting in AppConfigurationSettings)
+            {
+                appSettingsElement.Add(new XElement("add", new XAttribute("key", configSetting.Key), new XAttribute("value", configSetting.Value)));
             }
 
             // transport specification
@@ -341,7 +354,7 @@
                 }
 
                 Console.Out.Write(urlToMessageBody);
-
+       
                 return client.DownloadData(urlToMessageBody);
             }
         }
@@ -445,7 +458,15 @@
         ITransportIntegration transportToUse;
         public string PathToAppConfig
         {
-            get { return pathToAppConfig; }
+            get
+            {
+                if (String.IsNullOrWhiteSpace(pathToAppConfig))
+                {
+                    pathToAppConfig = Path.Combine(Path.GetTempPath(), Path.GetTempFileName());
+                    InitialiseAppConfig();
+                }
+                return pathToAppConfig;
+            }
         }
     }
 }
